añadir personaje movido a componente barra de acción
