+++ src/landing /Gallery.tsx
@@ -10,19 +10,10 @@ function Gallery() {
     const [characters, setCharacters] = useState<Character[]>([]);
     const [studentsToggle, setStudentsToggle] = useState(false);
     const [staffToggle, setStaffToggle] = useState(false)
-    const [showAddCharacter, setShowAddCharacter] = useState<boolean>(false);
 
     useEffect(() => {
         readCharacters()
     }, []);
-
-    function onAddCharacterClose() {
-        setShowAddCharacter(false)
-    }
-
-    function onAddCharacterOpen() {
-        setShowAddCharacter(true)
-    }
 
     async function readCharacters() {
         const res = await dataService.getCharacters();
@@ -57,8 +48,8 @@ function Gallery() {
 
     return (
         <>
-            <div className={`gallery ${showAddCharacter ? "invisible" : ""}`}>
-            {/* <div className='gallery'> */}
+            <div className='gallery'>
+                {/* <div className='gallery'> */}
                 <img className="namePage" src='./images/Harry_Potter.png' alt="HarryPotter" />
                 <h4 className='instructions'>Selecciona tu filtro</h4>
                 <div className='categoryButtonContainer'>
@@ -71,8 +62,7 @@ function Gallery() {
                     })}
                 </main>
             </div>
-            <ActionBar onClick={onAddCharacterOpen} />
-            {showAddCharacter ? <AddCharacter  onClose={onAddCharacterClose} onSave={readCharacters}/> : ''}
+            <ActionBar onSave={readCharacters} />
         </>
     );
 }
